requirements.txt
anyio==4.0.0
certifi==2023.7.22
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
pypdf==3.17.0
python-telegram-bot==20.6
sniffio==1.3.0
